Namespaces templates under -ember-table
This could have been done a number of ways but a subdirectory was used
as it allows us to still put templates in the global template namespace
if required.

'-ember-table' was chosen as a namespace. The - was added to prevent
accidental clashes between the app and library. It also seems to be
standard to prefix library/framework internal stuff. See the injection
of the active model adapter / serializer in ember itself